[CELEBORN-117] refactor CommitManager, implements M/R Partition CommitHandler to handle specific shuffle partition type.
