+++ client/src/main/scala/org/apache/celeborn/client/ChangePartitionManager.scala
@@ -271,7 +271,7 @@ class ChangePartitionManager(
       return
     }
 
-    if (lifecycleManager.commitManager.stageEndShuffleSet.contains(shuffleId)) {
+    if (lifecycleManager.commitManager.isStageEnd(shuffleId)) {
       logError(s"[handleChangePartition] shuffle $shuffleId already ended!")
       replyFailure(StatusCode.STAGE_ENDED)
       return
