+++ client/src/main/scala/org/apache/celeborn/client/CommitManager.scala
@@ -19,22 +19,21 @@ package org.apache.celeborn.client
 
 import java.util
 import java.util.concurrent.{ConcurrentHashMap, ScheduledExecutorService, ScheduledFuture, TimeUnit}
-import java.util.concurrent.atomic.{AtomicInteger, AtomicLong, LongAdder}
+import java.util.concurrent.atomic.{AtomicInteger, LongAdder}
 
 import scala.collection.JavaConverters._
 import scala.concurrent.duration.DurationInt
 
 import org.roaringbitmap.RoaringBitmap
 
+import org.apache.celeborn.client.CommitManager.CommittedPartitionInfo
+import org.apache.celeborn.client.commit.{CommitHandler, MapPartitionCommitHandler, ReducePartitionCommitHandler}
 import org.apache.celeborn.common.CelebornConf
 import org.apache.celeborn.common.internal.Logging
-import org.apache.celeborn.common.meta.{PartitionLocationInfo, WorkerInfo}
+import org.apache.celeborn.common.meta.WorkerInfo
 import org.apache.celeborn.common.protocol.{PartitionLocation, PartitionType, StorageInfo}
-import org.apache.celeborn.common.protocol.message.ControlMessages.{CommitFiles, CommitFilesResponse}
 import org.apache.celeborn.common.protocol.message.StatusCode
-import org.apache.celeborn.common.rpc.RpcEndpointRef
-import org.apache.celeborn.common.util.{ThreadUtils, Utils}
-import org.apache.celeborn.common.util.FunctionConverter._
+import org.apache.celeborn.common.util.ThreadUtils
 
 case class CommitPartitionRequest(
     applicationId: String,
@@ -55,16 +54,15 @@ case class ShuffleCommittedInfo(
     allInFlightCommitRequestNum: AtomicInteger,
     partitionInFlightCommitRequestNum: ConcurrentHashMap[Int, AtomicInteger])
 
+object CommitManager {
+  type CommittedPartitionInfo = ConcurrentHashMap[Int, ShuffleCommittedInfo]
+}
+
 class CommitManager(appId: String, val conf: CelebornConf, lifecycleManager: LifecycleManager)
   extends Logging {
+
   // shuffle id -> ShuffleCommittedInfo
-  private val committedPartitionInfo = new ConcurrentHashMap[Int, ShuffleCommittedInfo]()
-  val dataLostShuffleSet = ConcurrentHashMap.newKeySet[Int]()
-  val stageEndShuffleSet = ConcurrentHashMap.newKeySet[Int]()
-  private val inProcessStageEndShuffleSet = ConcurrentHashMap.newKeySet[Int]()
-  // shuffleId -> in processing partitionId set
-  private val inProcessMapPartitionEndIds = new ConcurrentHashMap[Int, util.Set[Int]]()
-  private val pushReplicateEnabled = conf.pushReplicateEnabled
+  private val committedPartitionInfo = new CommittedPartitionInfo
 
   private val batchHandleCommitPartitionEnabled = conf.batchHandleCommitPartitionEnabled
   private val batchHandleCommitPartitionExecutors = ThreadUtils.newDaemonCachedThreadPool(
@@ -80,12 +78,7 @@ class CommitManager(appId: String, val conf: CelebornConf, lifecycleManager: Lif
       None
     }
   private var batchHandleCommitPartition: Option[ScheduledFuture[_]] = _
-
-  private val totalWritten = new LongAdder
-  private val fileCount = new LongAdder
-
-  private val testRetryCommitFiles = conf.testRetryCommitFiles
-  private val commitEpoch = new AtomicLong()
+  private val commitHandlers = new ConcurrentHashMap[PartitionType, CommitHandler]()
 
   def start(): Unit = {
     batchHandleCommitPartition = batchHandleCommitPartitionSchedulerThread.map {
@@ -96,15 +89,7 @@ class CommitManager(appId: String, val conf: CelebornConf, lifecycleManager: Lif
               batchHandleCommitPartitionExecutors.submit {
                 new Runnable {
                   val partitionType = lifecycleManager.getPartitionType(shuffleId)
-                  def isPartitionInProcess(partitionId: Int): Boolean = {
-                    if (inProcessMapPartitionEndIds.containsKey(shuffleId) &&
-                      inProcessMapPartitionEndIds.get(shuffleId).contains(partitionId)) {
-                      true
-                    } else {
-                      false
-                    }
-                  }
-
+                  val commitHandler = getCommitHandler(shuffleId)
                   def incrementInflightNum(workerToRequests: Map[
                     WorkerInfo,
                     collection.Set[PartitionLocation]]): Unit = {
@@ -145,7 +130,8 @@ class CommitManager(appId: String, val conf: CelebornConf, lifecycleManager: Lif
                     if (partitionType == PartitionType.MAP) {
                       commitPartitionRequests.asScala.filterNot { request =>
                         shuffleCommittedInfo.handledCommitPartitionRequests
-                          .contains(request.partition) && isPartitionInProcess(
+                          .contains(request.partition) && commitHandler.isPartitionInProcess(
+                          shuffleId,
                           request.partition.getId)
                       }
                     } else {
@@ -163,8 +149,7 @@ class CommitManager(appId: String, val conf: CelebornConf, lifecycleManager: Lif
                       // partitions which are already committed by stageEnd process.
                       // But inProcessStageEndShuffleSet should have contain this shuffle id,
                       // can directly return.
-                      if (inProcessStageEndShuffleSet.contains(shuffleId) ||
-                        stageEndShuffleSet.contains(shuffleId)) {
+                      if (commitHandler.isStageEndOrInProcess(shuffleId)) {
                         logWarning(s"Shuffle $shuffleId ended or during processing stage end.")
                         shuffleCommittedInfo.commitPartitionRequests.clear()
                         Map.empty[WorkerInfo, Set[PartitionLocation]]
@@ -228,7 +213,7 @@ class CommitManager(appId: String, val conf: CelebornConf, lifecycleManager: Lif
                                 .toList
                                 .asJava
 
-                            commitFiles(
+                            commitHandler.commitFiles(
                               appId,
                               shuffleId,
                               shuffleCommittedInfo,
@@ -279,10 +264,7 @@ class CommitManager(appId: String, val conf: CelebornConf, lifecycleManager: Lif
 
   def removeExpiredShuffle(shuffleId: Int): Unit = {
     committedPartitionInfo.remove(shuffleId)
-    inProcessStageEndShuffleSet.remove(shuffleId)
-    inProcessMapPartitionEndIds.remove(shuffleId)
-    dataLostShuffleSet.remove(shuffleId)
-    stageEndShuffleSet.remove(shuffleId)
+    getCommitHandler(shuffleId).removeExpiredShuffle(shuffleId)
   }
 
   def registerCommitPartitionRequest(
@@ -299,401 +281,66 @@ class CommitManager(appId: String, val conf: CelebornConf, lifecycleManager: Lif
     }
   }
 
-  private def commitFiles(
-      applicationId: String,
-      shuffleId: Int,
-      shuffleCommittedInfo: ShuffleCommittedInfo,
-      worker: WorkerInfo,
-      masterIds: util.List[String],
-      slaveIds: util.List[String],
-      commitFilesFailedWorkers: ConcurrentHashMap[WorkerInfo, (StatusCode, Long)]): Unit = {
-
-    val res =
-      if (!testRetryCommitFiles) {
-        val commitFiles = CommitFiles(
-          applicationId,
-          shuffleId,
-          masterIds,
-          slaveIds,
-          lifecycleManager.shuffleMapperAttempts.get(shuffleId),
-          commitEpoch.incrementAndGet())
-        val res = requestCommitFilesWithRetry(worker.endpoint, commitFiles)
-
-        res.status match {
-          case StatusCode.SUCCESS => // do nothing
-          case StatusCode.PARTIAL_SUCCESS | StatusCode.SHUFFLE_NOT_REGISTERED | StatusCode.FAILED =>
-            logDebug(s"Request $commitFiles return ${res.status} for " +
-              s"${Utils.makeShuffleKey(applicationId, shuffleId)}")
-            commitFilesFailedWorkers.put(worker, (res.status, System.currentTimeMillis()))
-          case _ => // won't happen
-        }
-        res
-      } else {
-        // for test
-        val commitFiles1 = CommitFiles(
-          applicationId,
-          shuffleId,
-          masterIds.subList(0, masterIds.size() / 2),
-          slaveIds.subList(0, slaveIds.size() / 2),
-          lifecycleManager.shuffleMapperAttempts.get(shuffleId),
-          commitEpoch.incrementAndGet())
-        val res1 = requestCommitFilesWithRetry(worker.endpoint, commitFiles1)
-
-        val commitFiles = CommitFiles(
-          applicationId,
-          shuffleId,
-          masterIds.subList(masterIds.size() / 2, masterIds.size()),
-          slaveIds.subList(slaveIds.size() / 2, slaveIds.size()),
-          lifecycleManager.shuffleMapperAttempts.get(shuffleId),
-          commitEpoch.incrementAndGet())
-        val res2 = requestCommitFilesWithRetry(worker.endpoint, commitFiles)
-
-        res1.committedMasterStorageInfos.putAll(res2.committedMasterStorageInfos)
-        res1.committedSlaveStorageInfos.putAll(res2.committedSlaveStorageInfos)
-        res1.committedMapIdBitMap.putAll(res2.committedMapIdBitMap)
-        CommitFilesResponse(
-          status = if (res1.status == StatusCode.SUCCESS) res2.status else res1.status,
-          (res1.committedMasterIds.asScala ++ res2.committedMasterIds.asScala).toList.asJava,
-          (res1.committedSlaveIds.asScala ++ res1.committedSlaveIds.asScala).toList.asJava,
-          (res1.failedMasterIds.asScala ++ res1.failedMasterIds.asScala).toList.asJava,
-          (res1.failedSlaveIds.asScala ++ res2.failedSlaveIds.asScala).toList.asJava,
-          res1.committedMasterStorageInfos,
-          res1.committedSlaveStorageInfos,
-          res1.committedMapIdBitMap,
-          res1.totalWritten + res2.totalWritten,
-          res1.fileCount + res2.fileCount)
-      }
-
-    shuffleCommittedInfo.synchronized {
-      // record committed partitionIds
-      res.committedMasterIds.asScala.foreach({
-        case commitMasterId =>
-          val partitionUniqueIdList = shuffleCommittedInfo.committedMasterIds.computeIfAbsent(
-            Utils.splitPartitionLocationUniqueId(commitMasterId)._1,
-            (k: Int) => new util.ArrayList[String]())
-          partitionUniqueIdList.add(commitMasterId)
-      })
-
-      res.committedSlaveIds.asScala.foreach({
-        case commitSlaveId =>
-          val partitionUniqueIdList = shuffleCommittedInfo.committedSlaveIds.computeIfAbsent(
-            Utils.splitPartitionLocationUniqueId(commitSlaveId)._1,
-            (k: Int) => new util.ArrayList[String]())
-          partitionUniqueIdList.add(commitSlaveId)
-      })
-
-      // record committed partitions storage hint and disk hint
-      shuffleCommittedInfo.committedMasterStorageInfos.putAll(res.committedMasterStorageInfos)
-      shuffleCommittedInfo.committedSlaveStorageInfos.putAll(res.committedSlaveStorageInfos)
-
-      // record failed partitions
-      shuffleCommittedInfo.failedMasterPartitionIds.putAll(
-        res.failedMasterIds.asScala.map((_, worker)).toMap.asJava)
-      shuffleCommittedInfo.failedSlavePartitionIds.putAll(
-        res.failedSlaveIds.asScala.map((_, worker)).toMap.asJava)
-
-      shuffleCommittedInfo.committedMapIdBitmap.putAll(res.committedMapIdBitMap)
-
-      totalWritten.add(res.totalWritten)
-      fileCount.add(res.fileCount)
-      shuffleCommittedInfo.currentShuffleFileCount.add(res.fileCount)
-    }
+  def finalCommit(shuffleId: Int): Unit = {
+    getCommitHandler(shuffleId).finalCommit(
+      shuffleId,
+      r => lifecycleManager.recordWorkerFailure(r))
   }
 
-  private def requestCommitFilesWithRetry(
-      endpoint: RpcEndpointRef,
-      message: CommitFiles): CommitFilesResponse = {
-    val maxRetries = conf.requestCommitFilesMaxRetries
-    var retryTimes = 0
-    while (retryTimes < maxRetries) {
-      try {
-        if (testRetryCommitFiles && retryTimes < maxRetries - 1) {
-          endpoint.ask[CommitFilesResponse](message)
-          Thread.sleep(1000)
-          throw new Exception("Mock fail for CommitFiles")
-        } else {
-          return endpoint.askSync[CommitFilesResponse](message)
-        }
-      } catch {
-        case e: Throwable =>
-          retryTimes += 1
-          logError(
-            s"AskSync CommitFiles for ${message.shuffleId} failed (attempt $retryTimes/$maxRetries).",
-            e)
-      }
-    }
-
-    CommitFilesResponse(
-      StatusCode.FAILED,
-      List.empty.asJava,
-      List.empty.asJava,
-      message.masterIds,
-      message.slaveIds)
+  def finalPartitionCommit(shuffleId: Int, partitionId: Int): Boolean = {
+    getCommitHandler(shuffleId).finalPartitionCommit(
+      shuffleId,
+      partitionId,
+      r => lifecycleManager.recordWorkerFailure(r))
   }
 
-  def finalCommit(
-      applicationId: String,
-      shuffleId: Int,
-      fileGroups: ConcurrentHashMap[Integer, util.Set[PartitionLocation]]): Unit = {
-    if (stageEndShuffleSet.contains(shuffleId)) {
-      logInfo(s"[handleStageEnd] Shuffle $shuffleId already ended!")
-      return
-    }
-    inProcessStageEndShuffleSet.synchronized {
-      if (inProcessStageEndShuffleSet.contains(shuffleId)) {
-        logWarning(s"[handleStageEnd] Shuffle $shuffleId is in process!")
-        return
-      }
-      inProcessStageEndShuffleSet.add(shuffleId)
-    }
-    // ask allLocations workers holding partitions to commit files
-    val allocatedWorkers = lifecycleManager.shuffleAllocatedWorkers.get(shuffleId)
-    val dataLost = handleCommitFiles(applicationId, shuffleId, allocatedWorkers, None, fileGroups)
+  def isStageEnd(shuffleId: Int): Boolean = {
+    getCommitHandler(shuffleId).isStageEnd(shuffleId)
+  }
 
-    // reply
-    if (!dataLost) {
-      logInfo(s"Succeed to handle stageEnd for $shuffleId.")
-      // record in stageEndShuffleSet
-      stageEndShuffleSet.add(shuffleId)
+  def setStageEnd(shuffleId: Int): Unit = {
+    getCommitHandler(shuffleId).setStageEnd(shuffleId)
+  }
+
+  def isStageDataLost(shuffleId: Int): Boolean = {
+    getCommitHandler(shuffleId).isStageDataLost(shuffleId)
+  }
+
+  private def getCommitHandler(shuffleId: Int): CommitHandler = {
+    val partitionType = lifecycleManager.getPartitionType(shuffleId)
+    if (commitHandlers.containsKey(partitionType)) {
+      commitHandlers.get(partitionType)
     } else {
-      logError(s"Failed to handle stageEnd for $shuffleId, lost file!")
-      dataLostShuffleSet.add(shuffleId)
-      // record in stageEndShuffleSet
-      stageEndShuffleSet.add(shuffleId)
-    }
-    inProcessStageEndShuffleSet.remove(shuffleId)
-  }
-
-  private def handleCommitFiles(
-      applicationId: String,
-      shuffleId: Int,
-      allocatedWorkers: util.Map[WorkerInfo, PartitionLocationInfo],
-      partitionIdOpt: Option[Int] = None,
-      fileGroups: ConcurrentHashMap[Integer, util.Set[PartitionLocation]]): Boolean = {
-    val masterPartMap = new ConcurrentHashMap[String, PartitionLocation]
-    val slavePartMap = new ConcurrentHashMap[String, PartitionLocation]
-    val shuffleCommittedInfo = committedPartitionInfo.get(shuffleId)
-    val commitFilesFailedWorkers = new ConcurrentHashMap[WorkerInfo, (StatusCode, Long)]()
-    val commitFileStartTime = System.nanoTime()
-
-    val parallelism = Math.min(allocatedWorkers.size(), conf.rpcMaxParallelism)
-    ThreadUtils.parmap(
-      allocatedWorkers.asScala.to,
-      "CommitFiles",
-      parallelism) { case (worker, partitionLocationInfo) =>
-      if (partitionLocationInfo.containsShuffle(shuffleId.toString)) {
-        val masterParts =
-          partitionLocationInfo.getMasterLocations(shuffleId.toString, partitionIdOpt)
-        val slaveParts = partitionLocationInfo.getSlaveLocations(shuffleId.toString, partitionIdOpt)
-        masterParts.asScala.foreach { p =>
-          val partition = new PartitionLocation(p)
-          partition.setFetchPort(worker.fetchPort)
-          partition.setPeer(null)
-          masterPartMap.put(partition.getUniqueId, partition)
-        }
-        slaveParts.asScala.foreach { p =>
-          val partition = new PartitionLocation(p)
-          partition.setFetchPort(worker.fetchPort)
-          partition.setPeer(null)
-          slavePartMap.put(partition.getUniqueId, partition)
-        }
-
-        val (masterIds, slaveIds) = shuffleCommittedInfo.synchronized {
-          (
-            masterParts.asScala
-              .filterNot(shuffleCommittedInfo.handledCommitPartitionRequests.contains)
-              .map(_.getUniqueId).asJava,
-            slaveParts.asScala
-              .filterNot(shuffleCommittedInfo.handledCommitPartitionRequests.contains)
-              .map(_.getUniqueId).asJava)
-        }
-
-        commitFiles(
-          applicationId,
-          shuffleId,
-          shuffleCommittedInfo,
-          worker,
-          masterIds,
-          slaveIds,
-          commitFilesFailedWorkers)
-      }
-    }
-    lifecycleManager.recordWorkerFailure(commitFilesFailedWorkers)
-    // check all inflight request complete, for map partition, it's for single partitionId
-    waitInflightRequestComplete(shuffleId, shuffleCommittedInfo, partitionIdOpt)
-
-    // check all data lost or not, for map partition, it's for single partitionId
-    val dataLost = checkDataLost(applicationId, shuffleId, partitionIdOpt)
-
-    if (!dataLost) {
-      val committedPartitions = new util.HashMap[String, PartitionLocation]
-      getPartitionUniqueIds(shuffleCommittedInfo.committedMasterIds, partitionIdOpt).foreach { id =>
-        if (shuffleCommittedInfo.committedMasterStorageInfos.get(id) == null) {
-          logDebug(s"$applicationId-$shuffleId $id storage hint was not returned")
-        } else {
-          masterPartMap.get(id).setStorageInfo(
-            shuffleCommittedInfo.committedMasterStorageInfos.get(id))
-          masterPartMap.get(id).setMapIdBitMap(shuffleCommittedInfo.committedMapIdBitmap.get(id))
-          committedPartitions.put(id, masterPartMap.get(id))
-        }
-      }
-
-      getPartitionUniqueIds(shuffleCommittedInfo.committedSlaveIds, partitionIdOpt).foreach { id =>
-        val slavePartition = slavePartMap.get(id)
-        if (shuffleCommittedInfo.committedSlaveStorageInfos.get(id) == null) {
-          logDebug(s"$applicationId-$shuffleId $id storage hint was not returned")
-        } else {
-          slavePartition.setStorageInfo(shuffleCommittedInfo.committedSlaveStorageInfos.get(id))
-          val masterPartition = committedPartitions.get(id)
-          if (masterPartition ne null) {
-            masterPartition.setPeer(slavePartition)
-            slavePartition.setPeer(masterPartition)
-          } else {
-            logInfo(s"Shuffle $shuffleId partition $id: master lost, " +
-              s"use slave $slavePartition.")
-            slavePartition.setMapIdBitMap(shuffleCommittedInfo.committedMapIdBitmap.get(id))
-            committedPartitions.put(id, slavePartition)
+      commitHandlers.computeIfAbsent(
+        partitionType,
+        (partitionType: PartitionType) => {
+          partitionType match {
+            case PartitionType.REDUCE => new ReducePartitionCommitHandler(
+                appId,
+                conf,
+                lifecycleManager.shuffleAllocatedWorkers,
+                lifecycleManager.reducerFileGroupsMap,
+                committedPartitionInfo,
+                lifecycleManager.shuffleMapperAttempts)
+            case PartitionType.MAP => new MapPartitionCommitHandler(
+                appId,
+                conf,
+                lifecycleManager
+                  .shuffleAllocatedWorkers,
+                lifecycleManager.reducerFileGroupsMap,
+                committedPartitionInfo)
           }
-        }
-      }
-
-      committedPartitions.values().asScala.foreach { partition =>
-        val partitionLocations = fileGroups.computeIfAbsent(
-          partition.getId,
-          (k: Integer) => new util.HashSet[PartitionLocation]())
-        partitionLocations.add(partition)
-      }
-
-      logInfo(s"Shuffle $shuffleId " +
-        s"commit files complete. File count ${shuffleCommittedInfo.currentShuffleFileCount.sum()} " +
-        s"using ${(System.nanoTime() - commitFileStartTime) / 1000000} ms")
-    }
-
-    dataLost
-  }
-
-  private def getPartitionIds(
-      partitionIds: ConcurrentHashMap[String, WorkerInfo],
-      partitionIdOpt: Option[Int]): util.Map[String, WorkerInfo] = {
-    partitionIdOpt match {
-      case Some(partitionId) => partitionIds.asScala.filter(p =>
-          Utils.splitPartitionLocationUniqueId(p._1)._1 ==
-            partitionId).asJava
-      case None => partitionIds
+        })
     }
   }
 
-  private def waitInflightRequestComplete(
-      shuffleId: Int,
-      shuffleCommittedInfo: ShuffleCommittedInfo,
-      partitionIdOpt: Option[Int]): Unit = {
-    lifecycleManager.getPartitionType(shuffleId) match {
-      case PartitionType.REDUCE =>
-        while (shuffleCommittedInfo.allInFlightCommitRequestNum.get() > 0) {
-          Thread.sleep(1000)
-        }
-      case PartitionType.MAP => partitionIdOpt match {
-          case Some(partitionId) =>
-            if (shuffleCommittedInfo.partitionInFlightCommitRequestNum.containsKey(partitionId)) {
-              while (shuffleCommittedInfo.partitionInFlightCommitRequestNum.get(
-                  partitionId).get() > 0) {
-                Thread.sleep(1000)
-              }
-            }
-        }
-    }
+  def commitMetrics(): (Long, Long) = {
+    var totalWritten = 0L
+    var totalFileCount = 0L
+    commitHandlers.asScala.values.foreach(commitHandler => {
+      totalWritten += commitHandler.commitMetrics._1
+      totalFileCount += commitHandler.commitMetrics._2
+    })
+    (totalWritten, totalFileCount)
   }
-
-  private def checkDataLost(
-      applicationId: String,
-      shuffleId: Int,
-      partitionIdOpt: Option[Int]): Boolean = {
-    val shuffleCommittedInfo = committedPartitionInfo.get(shuffleId)
-    val shuffleKey = Utils.makeShuffleKey(applicationId, shuffleId)
-    val masterPartitionUniqueIdMap =
-      getPartitionIds(shuffleCommittedInfo.failedMasterPartitionIds, partitionIdOpt)
-    if (!pushReplicateEnabled && masterPartitionUniqueIdMap.size() != 0) {
-      val msg =
-        masterPartitionUniqueIdMap.asScala.map {
-          case (partitionUniqueId, workerInfo) =>
-            s"Lost partition $partitionUniqueId in worker [${workerInfo.readableAddress()}]"
-        }.mkString("\n")
-      logError(
-        s"""
-           |For shuffle $shuffleKey partition data lost:
-           |$msg
-           |""".stripMargin)
-      true
-    } else {
-      val slavePartitionUniqueIdMap =
-        getPartitionIds(shuffleCommittedInfo.failedSlavePartitionIds, partitionIdOpt)
-      val failedBothPartitionIdsToWorker = masterPartitionUniqueIdMap.asScala.flatMap {
-        case (partitionUniqueId, worker) =>
-          if (slavePartitionUniqueIdMap.asScala.contains(partitionUniqueId)) {
-            Some(partitionUniqueId -> (worker, slavePartitionUniqueIdMap.get(partitionUniqueId)))
-          } else {
-            None
-          }
-      }
-      if (failedBothPartitionIdsToWorker.nonEmpty) {
-        val msg = failedBothPartitionIdsToWorker.map {
-          case (partitionUniqueId, (masterWorker, slaveWorker)) =>
-            s"Lost partition $partitionUniqueId " +
-              s"in master worker [${masterWorker.readableAddress()}] and slave worker [$slaveWorker]"
-        }.mkString("\n")
-        logError(
-          s"""
-             |For shuffle $shuffleKey partition data lost:
-             |$msg
-             |""".stripMargin)
-        true
-      } else {
-        false
-      }
-    }
-  }
-
-  private def getPartitionUniqueIds(
-      ids: ConcurrentHashMap[Int, util.List[String]],
-      partitionIdOpt: Option[Int]): Iterable[String] = {
-    partitionIdOpt match {
-      case Some(partitionId) => ids.asScala.filter(_._1 == partitionId).flatMap(_._2.asScala)
-      case None => ids.asScala.flatMap(_._2.asScala)
-    }
-  }
-
-  def finalPartitionCommit(
-      applicationId: String,
-      shuffleId: Int,
-      fileGroups: ConcurrentHashMap[Integer, util.Set[PartitionLocation]],
-      partitionId: Int): Boolean = {
-    val inProcessingPartitionIds =
-      inProcessMapPartitionEndIds.computeIfAbsent(shuffleId, (k: Int) => new util.HashSet[Int]())
-    inProcessingPartitionIds.add(partitionId)
-
-    val allocatedWorkers =
-      lifecycleManager.shuffleAllocatedWorkers.get(shuffleId).asScala.filter(p =>
-        p._2.containsRelatedShuffleOrPartition(shuffleId.toString, Option(partitionId))).asJava
-
-    var dataCommitSuccess = true
-    if (!allocatedWorkers.isEmpty) {
-      dataCommitSuccess =
-        !handleCommitFiles(
-          applicationId,
-          shuffleId,
-          allocatedWorkers,
-          Option(partitionId),
-          fileGroups)
-    }
-
-    // release resources and clear worker info
-    allocatedWorkers.asScala.foreach { case (_, partitionLocationInfo) =>
-      partitionLocationInfo.removeAllRelatedPartitions(shuffleId.toString, Option(partitionId))
-    }
-    inProcessingPartitionIds.remove(partitionId)
-
-    dataCommitSuccess
-  }
-
-  def commitMetrics(): (Long, Long) = (totalWritten.sumThenReset(), fileCount.sumThenReset())
 }
